allowing selective execution for ramp-up and measurement-only runs

+++ Plugins/EASy-Producer/ScenariosTest/src/net/ssehub/easy/producer/scenario_tests/LocalRealTests.java
@@ -31,30 +31,31 @@ public class LocalRealTests extends RealTests {
         
         String projectName = "PL_WMS_Platform";
         File base = executeIndenicaCase(projectName, "0", "0");
-        Assert.assertNotNull(projectName + " execution failed, thus no base file available", base);
-
+        if (null != base) {
+            Assert.assertNotNull(projectName + " execution failed, thus no base file available", base);
+    
+            /*
+            String springConfigName = "Web.config";
+            String springConfigInstance = "WMSsolutionProduct/StorageBinSearchService/" + springConfigName;
+            assertFileEquality(new File(base, springConfigInstance),
+                new File(base, "expected_results/" + springConfigName));
+            
+            String dbCreationName = "1CreateDatabaseAndTables.sql";
+            String dbCreationInstance = "WMSsolutionProduct/Config_Resources/" + dbCreationName;
+            assertFileEquality(new File(base, dbCreationInstance),new File(base, "expected_results/" + dbCreationName));
+            
+            String dbDataName = "2InsertExampleData.sql";
+            String dbDataInstance = "WMSsolutionProduct/Config_Resources/" + dbDataName;
+            assertFileEquality(new File(base, dbDataInstance), new File(base, "expected_results/" + dbDataName));
+            
+            String orMapperName = "mapperfragment.xml";
+            String orMapperInstance = "WMSsolutionProduct/Config_Resources/" + orMapperName;
+            assertFileEquality(new File(base, orMapperInstance), new File(base, "expected_results/" + orMapperName));
+            
+            FileUtils.deleteQuietly(base);
+            */
+        }
         debug = false; // remove if working
-
-        /*
-        String springConfigName = "Web.config";
-        String springConfigInstance = "WMSsolutionProduct/StorageBinSearchService/" + springConfigName;
-        assertFileEquality(new File(base, springConfigInstance),
-            new File(base, "expected_results/" + springConfigName));
-        
-        String dbCreationName = "1CreateDatabaseAndTables.sql";
-        String dbCreationInstance = "WMSsolutionProduct/Config_Resources/" + dbCreationName;
-        assertFileEquality(new File(base, dbCreationInstance), new File(base, "expected_results/" + dbCreationName));
-        
-        String dbDataName = "2InsertExampleData.sql";
-        String dbDataInstance = "WMSsolutionProduct/Config_Resources/" + dbDataName;
-        assertFileEquality(new File(base, dbDataInstance), new File(base, "expected_results/" + dbDataName));
-        
-        String orMapperName = "mapperfragment.xml";
-        String orMapperInstance = "WMSsolutionProduct/Config_Resources/" + orMapperName;
-        assertFileEquality(new File(base, orMapperInstance), new File(base, "expected_results/" + orMapperName));
-        
-        FileUtils.deleteQuietly(base);
-        */
     }
 
     /**
@@ -65,8 +66,10 @@ public class LocalRealTests extends RealTests {
     @Test
     public void testQualiMasterApril14() throws IOException {
         String[] versions = {"0", "0"};
-        File base = executeCase("april14", versions, "QualiMaster/", null, true);
-        assertFileEqualityRec(new File(base, "expected"), base);
+        File base = executeCase("april14", versions, "QualiMaster/", null, Mode.REASON_INSTANTIATE);
+        if (null != base) {
+            assertFileEqualityRec(new File(base, "expected"), base);
+        }
     }
     
     /**
@@ -77,7 +80,7 @@ public class LocalRealTests extends RealTests {
      */
     @Test
     public void testQualiMasterFeb17() throws IOException {
-        testQualiMasterFeb17Impl();
+        testQualiMasterFeb17Impl(Mode.REASON_INSTANTIATE);
     }
     
     /**
@@ -88,7 +91,7 @@ public class LocalRealTests extends RealTests {
      */
     @Test
     public void testQualiMasterSep17() throws IOException {
-        testQualiMasterFeb17Impl();
+        testQualiMasterFeb17Impl(Mode.REASON_INSTANTIATE);
     }
 
 }
